Fixed possible CPU intensive go routine leak from watcher.
While investigating CPU with pprof on our services using this watcher, we found that massive CPU usage was cause by `runtime.siftdownTimer`.
The source of this large amount of generated Tickers was this place (time.Tick)

Moving that to time.After fixes the issue.
We cannot create Tick once and reuse that, since timeToSleep is dynamic.

See: https://forum.golangbridge.org/t/runtime-siftdowntimer-consuming-60-of-the-cpu/3773

Signed-off-by: Bartek Plotka <bwplotka@gmail.com>